requirements.txt
asgiref==3.8.1
Django==5.1.2
sqlparse==0.5.1
tzdata==2024.2
django-extensions==3.2.3
unidecode==1.3.8
markdown==3.7
Pillow==10.4.0
django-debug-toolbar==4.4.6
python-dotenv==1.0.1
python-telegram-bot==21.7
django-jazzmin==3.0.1
pylint==3.3.1
django-tinymce==4.1.0